backend/requirements.txt
fastapi
uvicorn
httpx
requests
pydantic
pytest
sqlalchemy
bcrypt
passlib
python-multipart
python-jose
python-dotenv
psycopg2-binary
jinja2